+++ Toggly.FeatureManagement.iOS/TogglySwiftUI/Sources/TogglySwiftUI.swift
@@ -5,4 +5,4 @@ import SwiftUI
 import TogglyCore
 
 /// Toggly SDK version for SwiftUI package
-public let togglySwiftUIVersion = "1.6.1"
+public let togglySwiftUIVersion = "1.7.0"