+++ Toggly.FeatureManagement.iOS/TogglyUIKit/Sources/TogglyUIKit.swift
@@ -6,5 +6,5 @@ import UIKit
 import TogglyCore
 
 /// Toggly SDK version for UIKit package
-public let togglyUIKitVersion = "1.6.1"
+public let togglyUIKitVersion = "1.7.0"
 #endif